+++ code/src/cljs/sixsq/nuvla/ui/edges/spec.cljs
@@ -53,5 +53,5 @@
    ::nuvlaboxes-in-clusters       nil
    ::nuvlabox-playbooks-cronjob   nil
    ::pagination                   (pagination-plugin/build-spec
-                                    :default-items-per-page 8)
+                                    :default-items-per-page 25)
    ::edges-search                 (full-text-search-plugin/build-spec)})