+++ code/src/cljs/sixsq/nuvla/ui/edges/views.cljs
@@ -32,7 +32,7 @@
     [sixsq.nuvla.ui.utils.zip :as zip]))
 
 
-(def view-type (r/atom :cards))
+(def view-type (r/atom :table))
 (def show-state-statistics (r/atom false))
 
 
@@ -764,7 +764,8 @@
     [pagination-plugin/Pagination
      {:db-path      [::spec/pagination]
       :change-event [::events/refresh-root]
-      :total-items  total-elements}]))
+      :total-items  total-elements
+      :i-per-page-multipliers [1 2 4]}]))
 
 
 (defn NuvlaboxTable